ClientSession improvements
Fleshed out the ClientSession state machine to follow the session state graph described by the IMAP specification.  ClientSession is now firmly a single-connection object that tracks the server state appropriately.  A higher-level class will need to be defined that manages multiple sessions (for when dealing with multiple folders, monitoring, etc.)