Now only prefetching emails required to maintain contiguous messages in database.
It's vital that the message table always have a contiguous list of messages from the highest (newest) to the oldest requested by the user.  If the user requests messages early enough that there would be a gap, the missing patch in the middle needs to be downloaded (UIDs, at least) to hold their position in the database.  Prior commit downloaded too many; this change makes sure only messages not being requested by user but unavailable in the database are prefetched.