Folder no longer attempts to download existing messages when re-selected: #3851
The SQLite Geary.Folder implementation now uses reference semantics, so when it's reselected
the same Folder object is used and in-memory state is consistent.  Also found an off-by-one
error when first querying an IMAP Folder that caused the most recent new email not to be
downloaded the first time.  Fixed as well in this commit.