Command tagging restructuring.
Commands were build with their commands prior to this change, for various reasons (mostly as
a convenience for myself).  This is not technically a good way to do things, it's far better
if the Tag is assigned right before the Command is sent to ensure they go out in order.