Removes spurious debug messages, better folder normalization
In addition to removing some debug messages that were probably
not as useful as they once were (and appeared a lot during normal
use of the program), also made slight change to normalizing the
folder when first opened ... although Eric's recent changes helped
with notifying messages when they were removed during this process,
I'm still finding the occasional old message listed in the folder
although it was archived earlier.  This change uses a more
thorough call to list_emails_by_id() to capture *all* changes
since the folder was opened.