Further engine rework: Revised interface to Geary.Folder
Now that the Geary.Folder interface is not tied to the internal
SQLite/IMAP folder objects, it is free to offer more data and
better information to the client via its signals.  This commit
makes the "email-removed" signal more useful and removes two
list_*() calls that were only useful to the EngineFolder itself
when manipulating the internal Folder objects.