configure wrapper script and top-level Makefile: closes #5143
This introduces a configure script that automates the basic task
of setting up a CMake build directory.  It also uses a top-level
Makefile (which the configure script generates, to prevent running
it until configure has executed) which does the build and copies
the final executables into the root of the source tree.