Database connection pooling
Reusing connections is a win with asynchronous transactions.  More
fine-tuning in the future of cache pages and such will help improve
Geary's database performance.

Also fixed a naming and namespace problem with Geary.DatabaseError
and a hole where, if a connection could not be opened (for resource
problems, most likely), the asynchronous transaction would never
complete.