Geary doesn't build with Vala 0.17.3: Closes #5561
Unfortunately, the problem is a binding issue that is not easily
worked around.  The temporary solution is to include a copy of the
Gdk.Pixbuf VAPI until this bug is closed and a patched VAPI
distributed: https://bugzilla.gnome.org/show_bug.cgi?id=680060

Note that with this commit, Geary now requires Vala 0.17.3 to build.