Move .desktop and GSettings schema file
Previous commit exposed a bug caused by moving some of the client's
resources in the source tree.  Since much of the client's resources
are stored off the source root directory, it doesn't make sense to
special case the .desktop file and the GSettings file and place them
in the client source itself.  This moves them out and makes the
appropriate changes in the CMake files and source code.