Select authentication method after STARTTLS: Closes #6019
The SMTP authentication method was being selected after initial
establishment of the session.  For the user's server, it was
reporting no AUTH methods available, and so PLAIN was selected.
However, after STARTTLS, the user's server reported only LOGIN,
causing the problem.

Smtp.ClientConnection now offers an establish_connection_async()
method which goes through all the steps to ensure quality-of-service
and security is set up.