Jump to furst unread message when selecting a conversation: Closes #6243
Although a duplicate of #4639, where we decided not to implement this
feature in lieu of implementing #5137 (compress collapsed read messages),
since we don't know when that will be ready, this is a good stop-gap,
as current behavior is annoying with long conversations.  Andrea's
code is cleanly delineated from the other code, and should be easily
removed if/when the time comes.