Fix issues rapidly selecting folders; fix #6318
We now use a small timer when you click a folder before we actually kick
off a network connection.  This effectively rate-limits how many
connections we make without significantly degrading the user experience.

Also we know that the folder selection function is not safely reentrant,
so that logic happens inside a mutex lock now.

The rate-limiting folder selection part is not an ideal solution, but
it's close to what we want (ideally, local results would be shown
immediately).  However, the mutex lock is just a bandaid over the larger
problem of how much global state we juggle in the controller.

Squashed commit of the following:

commit 44e46663bb9b513970e11a62428834668987bf33
Author: Charles Lindsay <chaz@yorba.org>
Date:   Wed Feb 6 17:29:22 2013 -0800

    Comment our mutex lock

commit d951204aa163c73ed60b51368fa900ad6414046c
Author: Charles Lindsay <chaz@yorba.org>
Date:   Wed Feb 6 16:59:32 2013 -0800

    Select folders only after a small timeout

commit 012097139bfbaa76967667b8fcdd324634cc2506
Author: Charles Lindsay <chaz@yorba.org>
Date:   Wed Feb 6 11:11:30 2013 -0800

    Fix reentrancy issues using mutex