Add delay for IMAP retries; fix #5357, fix #6263
There's now a minimum 15 second gap between IMAP retries.  This goes
easier on the CPU when the network is down or the user hasn't entered
their password.

Note that we special case the user's password changing to bypass the 15
second pause, so that they can immediately see their mail when they
(re-)enter their password.