Load all contacts into ContactStore: Closes #6921
All contacts are now loaded into the ContactStore, leaving it to
the application (i.e. client) to decide what to display and what
not to display.  This moves that responsibility into the composer's
autocompletion logic.

This leaves the door open for other contact-related code (i.e.
contact list regardless of their importance.