Fix spurious warning when creating toolbar
It looks like when Glade saved the toolbar.glade file recently, it
overwrote some apparently-necessary defaults.  The end result was that
one of our helpers was crashing because the gear menu button was getting
a null label.  If we explicitly set the label in the action, it seems to
make it happy.  This is preferably to undoing the change in the .glade
file because if we did that, the next time someone saved it, it'd just
break again.