Fix conv. filling when changing folders; fix #7099
Due to #5327, whenever you're on a folder with lots of conversations and
you're viewing one in the conversation viewer, if you then select a
different folder, you get many "conversation selected" signals.  This
fixes a bug where each one of those would trigger some background work
that wouldn't get cancelled, leading to many spurious messages showing
up in the conversation viewer as a result, instead of clearing the
viewer as expected.